Version 0.8.1

- Only sync FAT FS on TIM10 expire
- Attempt to reduce number of missed frames

+++ FATFS/App/fatfs.c
@@ -265,7 +265,7 @@ int FATFS_slwrite (FIL *fp, const void *buff, size_t len, int sync, int pos)
     }
   }
 
-  return bw;
+  return (int) bw;
 }
 
 int FATFS_close (FIL *fp)
